requirements.txt
streamlit>=1.32.0
anthropic>=0.39.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
srt>=3.5.3
